refactor(notification): move Notification into its own DDD package
Continues bringing the project to package-by-feature/DDD structure.
Notification, NotificationRepository, and NotificationType move from the
layer-based domain/ package into notification/, mirroring category/ and
task/. No behavior change — no service or controller existed for
Notification before this, so none is added here.
