+++ backend/src/main/java/br/com/calendar/notification/NotificationType.java
@@ -1,4 +1,4 @@
-package br.com.calendar.domain;
+package br.com.calendar.notification;
 
 
 public enum NotificationType {
